requirements.txt
# Web-Client dependencies
alembic==0.9.7
Babel==2.5.3
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Jinja2==2.10
Mako==1.0.7
PyJWT==1.5.3
python-dotenv==0.8.2
SQLAlchemy==1.2.1
WTForms==2.1

# Dev dependencies
autopep8==1.3.3
isort==4.2.15
jedi==0.11.1
pep8==1.7.1
pylint==1.8.1
